+++ app/Filament/Resources/LegacyChapters/Pages/ManageLegacyChapters.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\LegacyChapters\Pages;
+
+use App\Filament\Resources\LegacyChapters\LegacyChapterResource;
+use Filament\Resources\Pages\ManageRecords;
+
+class ManageLegacyChapters extends ManageRecords
+{
+    protected static string $resource = LegacyChapterResource::class;
+}