+++ app/Filament/Resources/LegacyLibraries/Pages/ManageLegacyLibraries.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\LegacyLibraries\Pages;
+
+use App\Filament\Resources\LegacyLibraries\LegacyLibraryResource;
+use Filament\Resources\Pages\ManageRecords;
+
+class ManageLegacyLibraries extends ManageRecords
+{
+    protected static string $resource = LegacyLibraryResource::class;
+}