+++ app/Filament/Resources/LegacyVerses/Pages/ManageLegacyVerses.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\LegacyVerses\Pages;
+
+use App\Filament\Resources\LegacyVerses\LegacyVerseResource;
+use Filament\Resources\Pages\ManageRecords;
+
+class ManageLegacyVerses extends ManageRecords
+{
+    protected static string $resource = LegacyVerseResource::class;
+}